+++ src/modules/osubot/db/index.js
@@ -1,3 +1,0 @@
-export userdb from './user'
-export statdb from './stat'
-export managedb from './manage'