+++ src/modules/osubot/index.js
@@ -1,3 +1,0 @@
-export * as commands from './commands'
-export inits from './init'
-export middlewares from './middleware'