+++ src/modules/osubot/map/index.js
@@ -1,3 +1,0 @@
-export { getDimensions, getDifficulty, getMods } from './difficulty'
-export { getPP, getFCPP } from './performance'
-export default from './aggregate'