+++ src/modules/utility/commands/index.js
@@ -1,4 +1,0 @@
-export roll from './roll'
-export test from './test'
-export batch from './batch'
-export announce from './announce'